Refactorizations: - Removed several unused variables - Use int instead of float instead of double where appropriate - Fixed keymapping in credits screen (u disables UI snow, not Tab)
git-svn-id: https://svn.code.sf.net/p/extremetuxracer/code/trunk@455 0420edf4-82e4-42fc-9478-35b55e6d67a3

+++ src/ogl.h
@@ -50,7 +50,7 @@ void PrintGLInfo();
 
 void set_material (const TColor& diffuse_colour,
                    const TColor& specular_colour,
-                   double specular_exp);
+                   float specular_exp);
 
 
 void PushRenderMode(TRenderMode mode);
